+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ### develop
+* correct argument name of UDP address & port
+  * `udp_receiver_sender_address` -> `udp_receiver_address`
+  * `udp_receiver_sender_port` -> `udp_receiver_port`
+  * should named "receiver"
 
 ### v0.1.1.431 (2023.8.8)